pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "embedding-pipeline"
version = "0.1.0"
description = "Image embedding pipeline using OpenCLIP LAION models"
requires-python = ">=3.10"
dependencies = [
    "open-clip-torch>=2.0.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-multipart>=0.0.6",
    "websockets>=11.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "playwright>=1.40.0",
    "httpx>=0.25.0",
]

[project.scripts]
embedding-pipeline = "embedding_pipeline.main:main"

[tool.setuptools.packages.find]
where = ["src"]
